requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
groq
scikit-learn
numpy
pandas
matplotlib
lightkurve
